导出优化

+++ kcylog/kcylog-system/src/main/java/com/kcylog/system/domain/SysReviewEmployee.java
@@ -92,7 +92,7 @@ public class SysReviewEmployee extends BaseEntity
     {
         return workTime;
     }
-    public void setWorkDay(Long workDay) 
+    public void setWorkDay(float workDay)
     {
         this.workDay = workDay;
     }
